sec: update parameter type in simple_crypto
Update descriptor buffer length parameter type required by RTA
descriptor routines (changed from uint16_t to unsigned to reflect
changes made in RTA library).

Reviewed-on: http://git.am.freescale.net:8181/1374
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>

+++ apps/simple_crypto/app/simple_crypto.c
@@ -433,7 +433,7 @@ static void *setup_init_descriptor(bool mode, struct test_param crypto_info)
 {
 	struct sec_descriptor_t *prehdr_desc;
 	uint32_t *shared_desc = NULL;
-	uint16_t shared_desc_len;
+	unsigned shared_desc_len;
 	int i, length;
 
 	prehdr_desc = __dma_mem_memalign(L1_CACHE_BYTES,
